Move the experimental await support to a different file
